HSEARCH-3046 Move HibernateOrmXClassOrdering to pojo spi
